Rename directory back from tharsis to internal/provider.
