+++ internal/provider/data_source_workspace_outputs_test.go
@@ -1,4 +1,4 @@
-package tharsis
+package provider
 
 import (
 	"os"
