+++ internal/provider/provider_test.go
@@ -1,4 +1,4 @@
-package tharsis
+package provider
 
 import (
 	"testing"
