+++ internal/provider/resource_managed_identities_test.go
@@ -1,4 +1,4 @@
-package tharsis
+package provider
 
 import (
 	"fmt"
