+++ internal/provider/resource_managed_identity_access_rules.go
@@ -1,4 +1,4 @@
-package tharsis
+package provider
 
 import (
 	"context"
