+++ internal/provider/version.go
@@ -1,4 +1,4 @@
-package tharsis
+package provider
 
 // Version is the current version of the provider, set via ldflags
 var Version = "dev"
